+++ preflight_checks.sh
@@ -1,0 +1,3 @@
+pub run test --platform vm
+dartanalyzer --fatal-infos .
+dartfmt -w lib test example